requirements.txt
transformers
pillow
torch
torchvision
matplotlib
pyarrow
google-cloud-vision
numpy==1.23.5
pandas==1.3.5
scikit-image
opencv-python==4.7.0.72
PyYAML
colorama
tqdm
requests
matplotlib
psutil
gunicorn
flask
scikit-learn
google-cloud-aiplatform
gunicorn
google-generativeai
gradio==4.44.1
thefuzz